refactor: move db types to a common location
