+++ mp/io/writer/__init__.py
@@ -1,0 +1,2 @@
+POSTGRESQL = 'postgresql'
+SQLITE = 'sqlite'
